Renamed globalMocks to utilMocks
