+++ client/test/unitTests/Routes/Library.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import Library from 'route/library/Library';
 import tabs from 'route/library/LibraryTabData';
-import { mockURLforLIB } from '../__mocks__/global_mocks';
+import { mockURLforLIB } from '../__mocks__/util_mocks';
 import {
   InitRouteTests,
   itDisplaysContentOfTabs,
