+++ client/test/unitTests/jest.setup.ts
@@ -1,5 +1,5 @@
 import '@testing-library/jest-dom';
-import './__mocks__/global_mocks';
+import './__mocks__/util_mocks';
 import './__mocks__/page_mocks';
 import './__mocks__/component_mocks';
